feat: admin 페이지 추가, 작가 전용 시나리오 작성 기능
- 전문가(EXPERT) 전용 MyPage에 시나리오 작성 폼 추가
  - case_summary, characters, clues, solution 입력 → JSON 자동 변환 후 저장
  - JSON 입력 오류 방지 및 사용자 친화적인 폼 기반 입력 제공
- ScenarioService에 DRAFT 기본 상태 저장 로직 반영
- AdminController에 시나리오 전체 조회, 승인, 반려, 삭제 기능 추가
- 시나리오 삭제 시 game_sessions 참조 무결성 문제 해결 (세션 삭제 후 시나리오 삭제 처리)
- 관리자 페이지(AdminDashboardPage)에서 시나리오 승인/반려/삭제 UI 및 액션 버튼 구현
- DB 스키마 조정: Scenario, GameResult, GameSession 엔티티 JSON 필드 구조 정리

+++ detective-backend/src/main/java/com/lingoguma/detective_backend/game/dto/GameResultRequest.java
@@ -6,9 +6,9 @@ import java.util.Map;
 @Getter @Setter
 @NoArgsConstructor @AllArgsConstructor @Builder
 public class GameResultRequest {
-    private Long sessionId;                 // ✅ game_results.session_id
-    private Long scenIdx;                   // ✅ game_results.scen_idx
-    private Long userIdx;                   // ✅ game_results.user_idx
+    private Integer sessionId;                 // ✅ game_results.session_id
+    private Integer scenIdx;                   // ✅ game_results.scen_idx
+    private Integer userIdx;                   // ✅ game_results.user_idx
     private Map<String, Object> answerJson; // JSON 그대로 받음
     private Map<String, Integer> skills;    // NLP 점수 JSON
     private boolean isCorrect;              // 정답 여부 (추가 컬럼)
